requirements.txt
astunparse==1.6.3
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1636541929935/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cryptography @ file:///croot/cryptography_1677533068310/work
ctcdecode @ file:///home/tbt/ctcdecode1/ctcdecode
cycler==0.11.0
dill==0.3.7
dlib==19.24.4
fonttools==4.38.0
ftfy==6.1.1
graphviz==0.20.1
idna @ file:///croot/idna_1666125576474/work
kiwisolver==1.4.5
lmdb==1.4.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.8.0.76
packaging==23.1
pandarallel==1.6.5
pandas==1.1.5
Pillow==9.4.0
psutil==5.9.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pynvml==11.5.3
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.4.16
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scipy==1.7.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
torch==1.13.1
torchjpeg==0.9.29
torchvision==0.14.1
tqdm==4.66.1
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///croot/urllib3_1673575502006/work
wcwidth==0.2.13
